requirements.txt
GitPython
pandas
matplotlib
openai
python-dotenv
pandas
matplotlib
scipy
seaborn
PyYAML
numpy
